make: test for stuff before linking & update bash

+++ Makefile
@@ -1,20 +1,17 @@
 VSCODECFGDIR := $(HOME)/Library/Application\ Support/Code/User
 
 init:
-	ln -s $(PWD)/hammerspoon $(HOME)/.hammerspoon
-	ln -s $(PWD)/emacs.d $(HOME)/.emacs.d
-	ln -s $(PWD)/zshrc.zsh $(HOME)/.zshrc
-	ln -s $(PWD)/vscode/settings.json $(VSCODECFGDIR)/settings.json
-	ln -s $(PWD)/vscode/keybindings.json $(VSCODECFGDIR)/keybindings.json
-	ln -s $(PWD)/vscode/projects.json $(VSCODECFGDIR)/projects.json
-	ln -s $(PWD)/vscode/snippets $(VSCODECFGDIR)/snippets
-	ln -s $(PWD)/vscode/theme-jellyfish $(HOME)/.vscode/extensions/chaseadamsio.vscode-theme-jellyfish
-	ln -s $(PWD)/ssh_config_common $(HOME)/.ssh/config_common
-	ln -s $(PWD)/gitconfig $(HOME)/.gitconfig
+	test -d $(HOME).hammerspoon || ln -s $(PWD)/hammerspoon $(HOME)/.hammerspoon
+	test -L $(HOME)/.bash_profile || ln -s $(PWD)/bashrc.bash $(HOME)/.bash_profile
+	test -L $(VSCODECFGDIR)/settings.json || ln -s $(PWD)/vscode/settings.json $(VSCODECFGDIR)/settings.json
+	test -L $(VSCODECFGDIR)/keybindings.json || ln -s $(PWD)/vscode/keybindings.json $(VSCODECFGDIR)/keybindings.json
+	test -L $(VSCODECFGDIR)/projects.json || ln -s $(PWD)/vscode/projects.json $(VSCODECFGDIR)/projects.json
+	test -d $(VSCODECFGDIR)/snippets || ln -s $(PWD)/vscode/snippets $(VSCODECFGDIR)/snippets
+	test -L $(HOME)/.ssh/config_commons || ln -s $(PWD)/ssh_config_common $(HOME)/.ssh/config_common
+	test -L $(HOME)/.gitconfig || ln -s $(PWD)/gitconfig $(HOME)/.gitconfig
 
 clean:
 	rm -rf $(HOME)/.hammerspoon
-	rm -rf $(HOME)/.emacs.d
 	rm -rf $(HOME)/.zshrc
 	rm -rf $(VSCODECFGDIR)/projects.json $(VSCODECFGDIR)/keybindings.json $(VSCODECFGDIR)/settings.json $(VSCODECFGDIR)/snippets
 	rm -rf $(HOME)/.ssh/config_common
